Fix: 부트페이 수정 - 조성윤 2025-02-16 18:06

+++ Prepay/app/src/main/java/com/example/prepay/data/response/StoreIdReq.kt
@@ -1,7 +1,7 @@
 package com.example.prepay.data.response
 
 data class StoreIdReq(
-    val latitude : Double,
-    val longitude : Double,
-    val teamId : Int
+    val latitude: Double,
+    val longitude: Double,
+    val teamId: Long
 )
